#5 6 game compare, this closes #5 and closes #6

+++ Quartett/app/src/main/java/de/in/uulm/map/quartett/gamesettings/GameSettingsContract.java
@@ -23,7 +23,7 @@ public interface GameSettingsContract {
 
         String getName();
 
-        int getLimit();
+        long getLimit();
 
         GameMode getMode();
